refactor: relocates media & card response types
