+++ packages/ai-chat/src/chat/AppShellStyles.scss
@@ -20,6 +20,7 @@
 @use "../globals/scss/chat-theme";
 @use "../globals/scss/chat-float-layout";
 @use "../globals/scss/chat-explainability-popover";
+@use "../globals/scss/max-width-utilities";
 
 @use "./components/imports" as components;
 @use "./components-legacy/imports" as legacyComponents;
