+++ packages/ai-chat/src/chat/components-legacy/MessageTypeComponent.tsx
@@ -34,11 +34,11 @@ import prefix from "@carbon/ai-chat-components/es/globals/settings.js";
 import { ResponseStopped } from "./ResponseStopped";
 import { SystemMessage } from "./SystemMessage";
 import { ConnectToHumanAgent } from "./responseTypes/humanAgent/ConnectToHumanAgent";
-import { AudioComponent } from "./responseTypes/audio/AudioComponent";
+import { AudioComponent } from "../components/messages/AudioComponent";
 import { ButtonItemComponent } from "./responseTypes/buttonItem/ButtonItemComponent";
-import { CardItemComponent } from "./responseTypes/card/CardItemComponent";
+import { CardItemComponent } from "../components/messages/CardItemComponent";
 import { PreviewCardComponent } from "./responseTypes/previewCard/PreviewCardComponent";
-import { CarouselItemComponent } from "./responseTypes/carousel/CarouselItemComponent";
+import { CarouselItemComponent } from "../components/messages/CarouselItemComponent";
 import { ConversationalSearch } from "./responseTypes/conversationalSearch/ConversationalSearch";
 import UserDefinedResponse from "./responseTypes/custom/UserDefinedResponse";
 import CustomFooterSlot from "./responseTypes/custom/CustomFooterSlot";
@@ -49,7 +49,7 @@ import { IFrameMessage } from "./responseTypes/iframe/IFrameMessage";
 import { Image } from "./responseTypes/image/Image";
 import { OptionComponent } from "./responseTypes/options/OptionComponent";
 import { StreamingRichText } from "./responseTypes/util/StreamingRichText";
-import { VideoComponent } from "./responseTypes/video/VideoComponent";
+import { VideoComponent } from "../components/messages/VideoComponent";
 import { useSelector } from "../hooks/useSelector";
 import { shallowEqual } from "../store/appStore";
 import { AppState } from "../../types/state/AppState";
