+++ packages/ai-chat/src/chat/components-legacy/responseTypes/_imports.scss
@@ -5,9 +5,7 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
-@use "maxWidth" as responseType-maxWidth;
 @use "humanAgent/imports" as responseType-agent;
-@use "card/imports" as responseType-card;
 @use "previewCard/imports" as responseType-previewCard;
 @use "carousel/imports" as responseType-carousel;
 @use "conversationalSearch/imports" as responseType-conversationalSearch;
