+++ packages/ai-chat/src/chat/components/_imports.scss
@@ -1,5 +1,5 @@
 /* 
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *  
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -10,3 +10,4 @@
 @use "modals/imports" as modals;
 @use "util/imports" as utils;
 @use "homeScreen/imports" as homescreen;
+@use "messages/imports" as messages;
