+++ packages/ai-chat/src/chat/components/messages/CardItemComponent.scss
