+++ packages/ai-chat/src/chat/components/messages/VideoComponent.tsx
@@ -9,12 +9,9 @@
 
 import React from "react";
 
-import { HasBaseHeight } from "../../../../types/utilities/HasBaseHeight";
-import {
-  MediaPlayer,
-  MediaPlayerContentConfig,
-} from "../../../components/util/MediaPlayer";
-import { MessageResponseTypes } from "../../../../types/messaging/Messages";
+import { HasBaseHeight } from "../../../types/utilities/HasBaseHeight.js";
+import { MediaPlayer, MediaPlayerContentConfig } from "../util/MediaPlayer.js";
+import { MessageResponseTypes } from "../../../types/messaging/Messages.js";
 
 type VideoComponentConfig = HasBaseHeight;
