+++ packages/ai-chat/src/chat/components/util/MediaPlayer.tsx
@@ -14,7 +14,7 @@ import React from "react";
 
 import { useSelector } from "../../hooks/useSelector";
 import { shallowEqual } from "../../store/appStore";
-import { VideoComponentConfig } from "../../components-legacy/responseTypes/video/VideoComponent";
+import { VideoComponentConfig } from "../messages/VideoComponent";
 import { MessageResponseTypes } from "../../../types/messaging/Messages";
 import { AppState } from "../../../types/state/AppState";
 import Card from "@carbon/ai-chat-components/es/react/card.js";
