+++ packages/ai-chat/src/globals/scss/_max-width-utilities.scss
@@ -1,0 +1,18 @@
+/* 
+ *  Copyright IBM Corp. 2025, 2026
+ *  
+ *  This source code is licensed under the Apache-2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ */
+
+.cds-aichat--max-width-small {
+  max-inline-size: 291px;
+}
+
+.cds-aichat--max-width-medium {
+  max-inline-size: 438px;
+}
+
+.cds-aichat--max-width-large {
+  max-inline-size: 585px;
+}
